+ Add ability to isotropic/anisotropic cell minimizations in LAMMPS-linked version + Fix small i/o bugs related to parallelization of LSQ code + Fix bug in MD "big" for loop definition
...

Notes: There may be a bug in the way Ewald-related calculations are done for the 
parallel version of the LSQ code. When the loop over frames is broken up over 
processors, I think the ability to tell if the cell lengths have changed is lost,
which can be problematic for Ewald calculations.